Added last two tests for checksum generation functions
The sha function was checked by a 3rd party alrogrythm.
The Blake function was harder to find and so was merely generated by this code.
If there is ever a better way to verify these results then I should update this
test. That being said, all the functions that rely on the blake checksum are
working and can be checked by 3rd party algorythims so I should be fine.