Made PoW generation behave when requesting the same hash twice.
Further work on the transaction manager to try and make it fault tolerant.
Added a simple refund function that will reverse the effect of a single send.
Added tracking information so that we can know what send has had an error (or
   has completed) in preparation for muti-send refunds.
Added timeouts to all of the transaction manager's wait periods.
Added "reverse-blacklist" to make sure that Address C and A never become
   associated.