Mixer can now supply funds to the transaction manager.
Added polling to transaction manager recieve confirmations. Whenever the mixer got involved it was almost always missing a few and the transaction would stall. My suspision is that this is becuase transactions of less than a certain amount don't get broadcast to the network. However it might also simply be because there were a ton of sends (~30).
Now if there are pending transactions for more than 10 seconds it starts manually polling to see if they are confirmed.
Maybe add polling to send confirmations as well?? Hasn't been a problem yet, but maybe it will someday.